refactor: fix url download using rwgps auth
