+++ .gitignore
@@ -3,6 +3,7 @@ files/
 outputs/
 .DS_Store
 .venv
+.env
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
