+++ .vscode/extensions.json
@@ -1,5 +1,6 @@
 {
     "recommendations": [
-        "streetsidesoftware.code-spell-checker"
+        "streetsidesoftware.code-spell-checker",
+        "charliermarsh.ruff"
     ]
 }
