+++ pyproject.toml
@@ -7,6 +7,7 @@ dependencies = [
     "requests<3.0.0,>=2.28.2",
     "typer>=0.16.0",
     "types-requests>=2.32.4.20250611",
+    "python-dotenv>=1.0.0",
 ]
 name = "ridewithgps-to-cuesheet"
 version = "0.4.0"
@@ -49,7 +50,7 @@ explicit_package_bases = true
 show_column_numbers = true
 
 [tool.poe.tasks]
-convert-route = "ridewithgps-to-cuesheet --url $URL --verbose"
+convert-route = "ridewithgps-to-cuesheet --verbose"
 lint = "ruff check ."
 format-check = "ruff format --check ."
 format = "ruff format ."
