--- requirements.txt
@@ -1,1 +1,0 @@
-poetry==2.1.3
