+++ sample.env
@@ -1,0 +1,2 @@
+RIDEWITHGPS_USERNAME=your_email_here
+RIDEWITHGPS_PASSWORD=your_password_here
