+++ src/ridewithgps_to_cuesheet/__init__.py
@@ -1,6 +1,7 @@
 """RideWithGPS to Cuesheet converter package."""
 
-from .ridewithgps import generate_excel
+from .conversion import generate_excel
+from .ridewithgps import AuthToken, authenticate, download_csv_content
 
-__version__ = "0.4.0"
-__all__ = ["generate_excel"]
+__version__ = "0.9.0"
+__all__ = ["generate_excel", "AuthToken", "authenticate", "download_csv_content"]
