+++ src/ridewithgps_to_cuesheet/logger.py
@@ -1,0 +1,4 @@
+
+import logging
+
+logger = logging.getLogger("ridewithgps-to-cuesheet")
